LPS-107518 Rename packages
