+++ modules/apps/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/internal/upgrade/AssetPublisherWebUpgrade.java
@@ -53,13 +53,11 @@ public class AssetPublisherWebUpgrade implements UpgradeStepRegistrator {
 
 		registry.register(
 			"1.0.2", "1.0.3",
-			new com.liferay.asset.publisher.web.internal.upgrade.v1_2_0.
-				UpgradePortletPreferences());
+			new com.liferay.asset.publisher.web.internal.upgrade.v1_0_3.UpgradePortletPreferences());
 
 		registry.register(
 			"1.0.3", "1.0.4",
-			new com.liferay.asset.publisher.web.internal.upgrade.v1_1_0.
-				UpgradePortletPreferences());
+			new com.liferay.asset.publisher.web.internal.upgrade.v1_0_4.UpgradePortletPreferences());
 	}
 
 	@Reference
